\name{cooccurExtra}
\alias{cooccurExtra}
\title{An extension use of R package "cooccur"}
\usage{
devtools::install_github("JotSoSerious/cooccurExtra")
}
\description{
This package is an extension use of R package "cooccur"
    Can be used for reproducing a huge work of cooccurence anaysis among a number of species pairs in different kinds of ingradients
    Some functional and visualize functions that may help doing the cooccurence anaysis
    In addition, some functions may help providing autometic summary tables
}
\examples{
library(cooccurExtra)
}

