\name{MatCOC}
\alias{cooccurExtra}
\title{
A function to create a matrix for the use of "cooccur" function.
}
\description{
This is a function of the package "cooccurExtra". The main idea of this function is creating a matrix for the use of "cooccur" function. The input object can be a data frame for the use of "CoocAna" or an output of "CoocAna". Either way the function will return a matrix for the use of "cooccur" function.
}
\usage{
MatCOC(OMat, ID.Col = "es", species = NA)
}
\arguments{
  \item{OMat}{
     A data frame, which contains species and ingradients information from different regions (for the use of "CoocAna" function). Or a output object of "CoocAna" function.
}
  \item{ID.Col}{
     Charater string, gives the name of a column in the data frame for the use of identifier. With a default value "es".
}
  \item{species}{
     A list of charaters with names of the needed species. Cannot be empty when the "OMat" is a data frame, otherwise the output will be "NULL".
}
}
\value{
  Returns a matrix in the form for the usage of "cooccur" function.
}
\author{
  Yingjia Jot He
}
\examples{
  # create a vector of charaters with all the needed species' names
  Four.Needed <- c("alpheus", "theora.lubrica",
                   "austrovenus.stutchburyi", "macomona.liliana")

  # ask for the model generated by "CoocAna" (with splits)
  ModelCA <- data(ModelCA)
  # the input object is a model build up with "CoocAna"
  MatCOC(OMat = ModelCA)[, 1:5]
  # the input object is a model build up with "CoocAna" with less species
  MatCOC(OMat = ModelCA, species = Four.Needed[-1])[, 1:5]

  # ask for the data frame to be analized
  macrofauna <- data(macrofauna)
  # the input "OMat" is a data frame for "CoocAna" function
  MatCOC(OMat = macrofauna, species = Four.Needed)[, 1:5]
}
