% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-mpdata.R
\docType{data}
\name{myData_1}
\alias{myData_1}
\title{The needed data frames in my master project.}
\format{Data frame which contains information of species and ingradients.}
\usage{
myData_1
}
\description{
The needed data frames in my master project.
}
\examples{
data(myData_1)
}
\keyword{datasets}
