% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_htmlwidget.R
\name{plot_htmlwidget}
\alias{plot_htmlwidget}
\title{Export cooccurence tables as an image.}
\usage{
plot_htmlwidget(widget, file = NULL)
}
\arguments{
\item{widget}{\code{htmlwidget} object.}

\item{file}{The name of the file where the data will be saved. Default to \code{NULL}, no saving required.}
}
\description{
This function allow users to display cooccurence tables from the \code{cooccurExtra} in a word or pdf document.
Since by default htmlwidget objects cannot be displayed in word or pdf documents.
This function also allow users to save these tables as a PNG locally.
}
\details{
To render the htmlwidget into an image. A headless browser is used to capture it, to do this the Chrome
Dev Tool is used, the \code{chromote} package provides this functionality.
}
\examples{
# require packages
# devtools::install_github("rstudio/chromote")
# install.packages("showimage")
library(chromote)
library(showimage)
# ask for a "coocana" output model
data(ModelCA)
mytest = displaytable(mymod = modelca)
plot_htmlwidget(mytest[[3]])
plot_htmlwidget(mytest[[4]])
}
\author{
Yingjia J He, Charco Hui
}
