\name{snineeded}
\alias{snineeded}
\docType{data}
\title{
The needed species and ingradients in my master project.
}
\description{
The needed species and ingradients in my master project.
}
\usage{data("snineeded")}
\format{
  Two vectors with 10 species and 3 ingradients.
  \describe{
    \item{Tenspecies}{A vector holds the names of ten species.}
    \item{ThreeIngradients}{A vector holds the names of three ingradients.}
  }
}
\examples{
data("snineeded")
}
