\name{CheckPNR}
\alias{cooccurExtra}
\title{
A function to distinct the cooccurence of species pairs from a "cooccur" ouput model object.
}
\description{
This is a function of the package "cooccurExtra". The main idea of this function is distincting the cooccurence ("positive", "negative", or "random") of species pairs from a "cooccur" ouput model object. It provides an output of table in the class of data frame to show probabilities as well as the classifying positive, negative or no cooccurence associations between species in each pair.
}
\usage{
CheckPNR(cooccur.mod)
}
\arguments{
  \item{cooccur.mod}{
     An model object generated by "cooccur" function from package "cooccur".
}
}
\value{
  Returns a data frame with all the needed pairs of species as well as the p-values and the cooccurence between the species in each pair.
  \item{sp1_name}{
     Name of a specie in a pair.
}
  \item{sp2_name}{
     Name of the other specie in a pair.
}
  \item{p_gt}{
     Probabilities for rejecting classifying positive and negative associations between the species in each pairs.
}
  \item{PNR}{
     The cooccurence associations ("positive", "negative", or "random") between the species in each pairs.
}
}
\author{
  Yingjia Jot He
}
\examples{
  # ask for the model generated by "CoocAna" (with splits)
  ModelCA <- data(ModelCA)
  ModelCOC <- ModelCA$Model$`Full Data`

  # the default with two kinds of plots
  mytest <- CheckPNR(ModelCOC)

  # show the output
  mytest
}
